Update hash.inl
Line 274 is definitely a problem, someone must have had thought `key_blob` was a reference to the public key blob, which is, in fact, a full-blown pointer and hence created a double reference to the key blob. This is also related to key_blob.

Also there's no need to statically expand the key size since the compiler will cache the value of key_blob->size and access it instead.